refactor: clean up schemas and types (#10)
