--- src/services/indexing/indexing.service.ts
@@ -1,234 +1,0 @@
-import { randomUUID } from "crypto";
-import {
-  MongoResource,
-  QdrantPoint,
-  MemgraphNode,
-  generateMongoId,
-  IndexRequest,
-  IndexResponse,
-} from "../../types/index.js";
-import { DocumentStore } from "../../stores/document.store.js";
-import { VectorStore } from "../../stores/vector.store.js";
-import { GraphStore } from "../../stores/graph.store.js";
-import { ChunkerService } from "./chunker.js";
-import { EmbedderService } from "./embedder.js";
-
-/**
- * Main indexing service - Orchestrates the entire indexing pipeline
- * Single-tenant version (no workspace isolation)
- */
-export class IndexingService {
-  constructor(
-    private documentStore: DocumentStore,
-    private vectorStore: VectorStore,
-    private graphStore: GraphStore,
-    private chunker: ChunkerService,
-    private embedder: EmbedderService
-  ) {}
-
-  /**
-   * Index a single request
-   */
-  async index(request: IndexRequest): Promise<IndexResponse> {
-    const jobId = randomUUID();
-    const startTime = Date.now();
-
-    try {
-      console.log(`[${jobId}] Starting indexing`);
-
-      // Step 1: Extract resources from MCP tool result
-      const extractedResources = await this.chunker.extractFromIndexRequest(
-        request
-      );
-
-      console.log(
-        `[${jobId}] Extracted ${extractedResources.length} resources`
-      );
-
-      // Step 2: Process each resource
-      let indexed = 0;
-      let failed = 0;
-
-      for (const extracted of extractedResources) {
-        try {
-          await this.processResource(extracted, request.source.type);
-          indexed++;
-        } catch (error) {
-          console.error(
-            `[${jobId}] Failed to process resource ${extracted.id}:`,
-            error
-          );
-          failed++;
-        }
-      }
-
-      const durationMs = Date.now() - startTime;
-
-      console.log(
-        `[${jobId}] Completed: ${indexed} indexed, ${failed} failed in ${durationMs}ms`
-      );
-
-      return {
-        jobId,
-        status: "completed",
-        stats: {
-          resourcesProcessed: extractedResources.length,
-          resourcesIndexed: indexed,
-          resourcesFailed: failed,
-          durationMs,
-        },
-      };
-    } catch (error) {
-      console.error(`[${jobId}] Indexing job failed:`, error);
-
-      return {
-        jobId,
-        status: "failed",
-        error: error instanceof Error ? error.message : String(error),
-      };
-    }
-  }
-
-  /**
-   * Process a single extracted resource
-   */
-  private async processResource(
-    extracted: any,
-    sourceType: string
-  ): Promise<void> {
-    const mongoId = generateMongoId(sourceType as any, extracted.resourceId);
-
-    // Check if resource needs chunking
-    const needsChunking = extracted.textContent.length > 2000;
-
-    if (needsChunking) {
-      // Process with chunking
-      await this.processWithChunking(extracted, mongoId);
-    } else {
-      // Process without chunking (single document)
-      await this.processSingleDocument(extracted, mongoId);
-    }
-
-    // Create graph node
-    const node: MemgraphNode = {
-      label: "", // Will be set by store based on type
-      id: mongoId,
-      type: extracted.type,
-      title: extracted.title,
-    };
-
-    await this.graphStore.createNode(node);
-
-    // Create relationships if any
-    if (extracted.relationships && extracted.relationships.length > 0) {
-      await this.graphStore.createRelationships(extracted.relationships);
-    }
-  }
-
-  /**
-   * Process document without chunking
-   */
-  private async processSingleDocument(
-    extracted: any,
-    mongoId: string
-  ): Promise<void> {
-    // Generate embedding
-    const vector = await this.embedder.embedWithPreprocessing(
-      extracted.textContent
-    );
-
-    // Create Qdrant point
-    const qdrantId = randomUUID();
-    const qdrantPoint: QdrantPoint = {
-      id: qdrantId,
-      vector,
-      payload: {
-        mongoId,
-      },
-    };
-
-    // Create MongoDB resource
-    const mongoResource: MongoResource = {
-      _id: mongoId,
-      source: extracted.source,
-      resourceId: extracted.resourceId,
-      type: extracted.type,
-      title: extracted.title,
-      textContent: extracted.textContent,
-      people: extracted.people,
-      primaryDate: extracted.primaryDate,
-      attributes: extracted.attributes,
-      rawData: extracted.rawData,
-      qdrantIds: [qdrantId],
-      embeddingVersion: 1,
-      indexedAt: new Date(),
-      updatedAt: new Date(),
-    };
-
-    // Save to all databases
-    await Promise.all([
-      this.documentStore.save(mongoResource),
-      this.vectorStore.upsertPoints([qdrantPoint]),
-    ]);
-  }
-
-  /**
-   * Process document with chunking
-   */
-  private async processWithChunking(
-    extracted: any,
-    mongoId: string
-  ): Promise<void> {
-    // Chunk the text
-    const chunks = this.chunker.chunkText(extracted.textContent);
-
-    console.log(`Chunking document ${mongoId} into ${chunks.length} chunks`);
-
-    // Generate embeddings for all chunks
-    const chunkTexts = chunks.map((c) => c.text);
-    const vectors = await this.embedder.embedBatchWithPreprocessing(chunkTexts);
-
-    // Create Qdrant points for each chunk
-    const qdrantPoints: QdrantPoint[] = chunks.map((chunk, idx) => ({
-      id: randomUUID(),
-      vector: vectors[idx],
-      payload: {
-        mongoId,
-        chunkIndex: chunk.index,
-        chunkStart: chunk.start,
-        chunkEnd: chunk.end,
-      },
-    }));
-
-    // Create MongoDB resource
-    const mongoResource: MongoResource = {
-      _id: mongoId,
-      source: extracted.source,
-      resourceId: extracted.resourceId,
-      type: extracted.type,
-      title: extracted.title,
-      textContent: extracted.textContent, // Store full content
-      people: extracted.people,
-      primaryDate: extracted.primaryDate,
-      attributes: extracted.attributes,
-      rawData: extracted.rawData,
-      qdrantIds: qdrantPoints.map((p) => p.id),
-      embeddingVersion: 1,
-      indexedAt: new Date(),
-      updatedAt: new Date(),
-    };
-
-    // Save to all databases
-    await Promise.all([
-      this.documentStore.save(mongoResource),
-      this.vectorStore.upsertPoints(qdrantPoints),
-    ]);
-  }
-
-  /**
-   * Batch index multiple requests
-   */
-  async indexBatch(requests: IndexRequest[]): Promise<IndexResponse[]> {
-    return Promise.all(requests.map((req) => this.index(req)));
-  }
-}
